removed trainer support
