+++ fluence/datasets/__init__.py
@@ -1,1 +1,1 @@
-from .siamese_dataset import SiameseGlueDataset, siamese_data_collator
+from .meta_dataset import MetaDataset
