+++ fluence/models/__init__.py
@@ -1,2 +1,2 @@
 from .orthogonal import HEXProjection, OrthogonalTransformer
-from .siamese_model import SiameseTransformer, SiameseTransformerAdd
+from .siamese_model import SiameseTransformer
