--- fluence/utils/__init__.py
@@ -1,1 +1,0 @@
-from .siamese_utils import SiameseModelArguments, SiameseTrainer
